feat(dataviews): Implement TableDataviewRenderer for enhanced table rendering
- Added TableDataviewRenderer class to handle table rendering with sorting and pagination.
- Introduced methods for sorting and paginating querysets based on user preferences.
- Updated model definitions to support new dataview types and configurations.
- Refactored user list view preferences to integrate with new dataview options.
- Enhanced CSS styles for better UI consistency across dataviews.
- Updated tests to cover new functionality in pivot and kanban dataviews.
- Created a base enum definition registry for future extensibility.
