+++ bloomerp/django_bloomerp/bloomerp/components/objects/dataviews/dataview.py
@@ -5,6 +5,7 @@ from django.shortcuts import render
 from django.urls import reverse
 from django.core.exceptions import FieldDoesNotExist
 from bloomerp.components.application_fields.filters import filters_init
+from bloomerp.dataviews.registry import DataviewType
 from bloomerp.models.definition import ObjectAction, get_model_config
 from bloomerp.permissions.definition import BloomerpPermission
 from bloomerp.permissions.manager import UserPolicyManager
@@ -18,7 +19,7 @@ from django.contrib.contenttypes.models import ContentType
 from bloomerp.services.user_services import get_data_view_fields
 from bloomerp.services.object_services import string_search_on_queryset
 from bloomerp.utils.filters import filter_model
-from bloomerp.models.users.user_list_view_preference import UserListViewPreference, DataviewType
+from bloomerp.models.users.user_list_view_preference import UserListViewPreference
 from bloomerp.models import ApplicationField
 from django.db.models import Model, QuerySet
 from dataclasses import dataclass
@@ -223,7 +224,9 @@ def _get_actions(model:type[Model]) -> list[ObjectAction]:
     return []
 
 
-def _get_dataview_type_definition(view_type: str) -> DataviewTypeDefinition:
+def _get_dataview_type_definition(
+    view_type: str,
+) -> DataviewTypeDefinition | None:
     try:
         return DataviewType.from_key(view_type)
     except ValueError:
