+++ bloomerp/django_bloomerp/bloomerp/dataviews/calendar/__init__.py
@@ -1,0 +1,1 @@
+"""Calendar dataview capability."""
