+++ bloomerp/django_bloomerp/bloomerp/dataviews/card/__init__.py
@@ -1,0 +1,1 @@
+"""Card dataview capability."""
