+++ bloomerp/django_bloomerp/bloomerp/dataviews/gant/__init__.py
@@ -1,0 +1,1 @@
+"""Gantt dataview capability."""
