+++ bloomerp/django_bloomerp/bloomerp/dataviews/kanban/__init__.py
@@ -1,0 +1,1 @@
+"""Kanban dataview capability."""
