+++ bloomerp/django_bloomerp/bloomerp/dataviews/pivot_table/__init__.py
@@ -1,0 +1,1 @@
+"""Pivot-table dataview capability."""
