+++ bloomerp/django_bloomerp/bloomerp/dataviews/table/__init__.py
@@ -1,0 +1,1 @@
+"""Table dataview capability."""
