+++ bloomerp/django_bloomerp/bloomerp/dataviews/table/config.py
@@ -1,0 +1,83 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any, Literal
+
+from django import forms
+from django.db.models import QuerySet
+from django.utils.translation import gettext_lazy as _
+
+from bloomerp.dataviews.base import (
+    BaseDataView,
+    PageSize,
+    PreferenceOption,
+    application_field_name_choices,
+    page_size_choices,
+)
+
+if TYPE_CHECKING:
+    from bloomerp.models.application_field import ApplicationField
+
+
+class TableDataView(BaseDataView):
+    """A declarative table dataview."""
+
+    view_type: Literal["table"] = "table"
+    page_size: Literal[10, 25, 50, 100] = 25
+    sort_field: str | None = None
+    sort_direction: Literal["asc", "desc"] = "asc"
+
+
+def sort_field_choices(
+    application_fields: QuerySet[ApplicationField],
+) -> dict[str, Any]:
+    return {
+        "choices": application_field_name_choices(
+            application_fields,
+            include_empty=True,
+            empty_label=_("Default"),
+        ),
+        "coerce": lambda value: value or None,
+        "empty_value": None,
+    }
+
+
+def sort_direction_choices(
+    _application_fields: QuerySet[ApplicationField],
+) -> dict[str, Any]:
+    return {
+        "choices": [
+            ("asc", _("Ascending")),
+            ("desc", _("Descending")),
+        ]
+    }
+
+
+TABLE_OPTIONS = [
+    PreferenceOption(
+        key="page_size",
+        label=_("Page size"),
+        field_cls=forms.TypedChoiceField,
+        field_attrs_func=page_size_choices,
+        description=_("The number of records shown on each page."),
+        data_type=int,
+        default_value=PageSize.SIZE_25,
+    ),
+    PreferenceOption(
+        key="sort_field",
+        label=_("Sort on"),
+        field_cls=forms.TypedChoiceField,
+        field_attrs_func=sort_field_choices,
+        description=_("The field used for table sorting."),
+        data_type=str | None,
+        default_value=None,
+    ),
+    PreferenceOption(
+        key="sort_direction",
+        label=_("Sort direction"),
+        field_cls=forms.ChoiceField,
+        field_attrs_func=sort_direction_choices,
+        description=_("The direction used for table sorting."),
+        data_type=str,
+        default_value="asc",
+    ),
+]
