+++ bloomerp/django_bloomerp/bloomerp/dataviews/table/renderer.py
@@ -4,8 +4,7 @@ from django.core.exceptions import FieldDoesNotExist
 from django.db.models import QuerySet
 from django.http import HttpRequest
 
-from .base import DataviewPagination
-from .base import BaseDataviewRenderer
+from ..base import BaseDataviewRenderer, DataviewPagination
 
 
 SORT_DIRECTIONS = {"asc", "desc"}
