+++ bloomerp/django_bloomerp/bloomerp/models/users/__init__.py
@@ -1,6 +1,6 @@
 from .user import AbstractBloomerpUser, User
 from .bookmark import Bookmark
-from .user_list_view_preference import UserListViewPreference, DataviewType
+from .user_list_view_preference import UserListViewPreference
 from .user_object_layout_preference import UserObjectLayoutPreference
 from .user_detail_view_tabs_preference import (
     UserDetailViewTabItem,
@@ -12,7 +12,6 @@ __all__ = [
     'User',
     'Bookmark',
     'UserListViewPreference',
-    'DataviewType',
     'UserObjectLayoutPreference',
     'UserDetailViewTabItem',
     'UserDetailViewTabsPreference',
