+++ bloomerp/django_bloomerp/bloomerp/utils/registry.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class BaseEnumDefinitionRegistry(Enum):
+    _id_field : str = "key"
+    _model = None
+    
+    
+
